Tests/LibWeb/Text/expected/css/disabled-style-sheet.txt

Before disabling style sheet: rgb(255, 255, 0)
After disabling style sheet: rgba(0, 0, 0, 0)
After explicitly enabling linked style sheet: rgb(0, 0, 255)
